move same implementations to util package
